Fix #474. Make helper for redefined devise methods
