+++ app/controllers/application_controller.rb
@@ -19,24 +19,6 @@ class ApplicationController < ActionController::Base
   helper_method :current_or_guest_user
   helper_method :f
 
-  def signed_in_root_path(resource_or_scope)
-    scope = Devise::Mapping.find_scope!(resource_or_scope)
-    router_name = Devise.mappings[scope].router_name
-
-    home_path = "#{scope}_root_path"
-
-    context = router_name ? send(router_name) : self
-    if context.respond_to?(home_path, true)
-      context.send(home_path)
-    elsif context.respond_to?(:root_path)
-      context.root_path(locale:)
-    elsif respond_to?(:root_path)
-      root_path(locale:)
-    else
-      "/#{locale}"
-    end
-  end
-
   private
 
   def redirect_root_domain
