+++ app/controllers/web/devise/passwords_controller.rb
@@ -2,4 +2,5 @@
 
 class Web::Devise::PasswordsController < Devise::PasswordsController
   include LocaleConcern
+  include DeviseHelper
 end
