+++ app/controllers/web/omniauth_callbacks_controller.rb
@@ -2,6 +2,7 @@
 
 class Web::OmniauthCallbacksController < Devise::OmniauthCallbacksController
   include LocaleConcern
+  include DeviseHelper
 
   def github
     @user = User.from_omniauth(request.env['omniauth.auth'])
